fix(cards): bottom-bar strip spans full card width + homescreen support + rename
User feedback on the prior strip implementation:

1. Setting renamed from "Card layout" to "Rating layout". The option only
   controls where the rating appears, not the overall card structure --
   the prior name was misleading. Choice labels are now "On right"
   (default pill column) and "Bottom bar" (tier-colored strip).

2. The strip was confined to game-card-body width (right of the
   thumbnail). renderGameCard now wraps thumb / body / right in a new
   game-card-row, and the strip becomes a sibling of the row. game-card
   itself becomes flex-column with overflow: hidden so the strip extends
   the full card width including under the thumbnail, with rounded
   corners respected at the bottom.

3. The homescreen (index.html) uses its own pgCardHtml / .pg-card markup
   instead of renderGameCard, so the prior change didn't reach it. Same
   restructure applied to the homescreen card: pg-card-row wraps the
   inner row, pg-card-strip becomes a sibling, CSS in css/index/index.css
   gets the matching data-card-layout and tier-color rules.

The bar matches ProtonDB's design now: full card width, tier-colored
gradient background, prominent tier name on the left, store pill chip
on the right.

Tests: card.test.js asserts the strip is a sibling of the row (outside
game-card-body), so future refactors don't accidentally re-nest it.
