+++ auth.html
@@ -11,7 +11,7 @@
 <link rel="stylesheet" href="css/shared/base.css?v=45c7e06d">
 <link rel="stylesheet" href="css/shared/topbar.css?v=f4f0510d">
 <link rel="stylesheet" href="css/shared/site.css?v=4d518a51">
-<link rel="stylesheet" href="css/shared/cards.css?v=f95cc608">
+<link rel="stylesheet" href="css/shared/cards.css?v=5a77b933">
 <link rel="stylesheet" href="css/auth/auth.css?v=ca512b85">
 </head>
 <body>
