+++ css/index/index.css
@@ -235,15 +235,49 @@
 }
 .pg-card {
   display: flex;
-  align-items: center;
-  gap: 14px;
-  padding: 8px 14px 8px 8px;
+  flex-direction: column;
   border: 1px solid var(--border);
   border-radius: 10px;
   background: var(--card);
+  overflow: hidden;
   text-decoration: none;
   transition: border-color 0.15s, background 0.15s, transform 0.05s;
 }
+.pg-card-row {
+  display: flex;
+  align-items: center;
+  gap: 14px;
+  padding: 8px 14px 8px 8px;
+}
+/* Hide the bottom strip by default; "Rating layout: Bottom bar" flips it on. */
+.pg-card-strip { display: none; }
+[data-card-layout="strip"] .pg-card-strip {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 8px;
+  padding: 7px 14px;
+  background: var(--s2);
+  color: var(--muted);
+}
+[data-card-layout="strip"] .pg-card-strip[data-tier="platinum"] { background: linear-gradient(90deg, #b4c7dc, #cad8e7); color: #0a0c10; }
+[data-card-layout="strip"] .pg-card-strip[data-tier="gold"]     { background: linear-gradient(90deg, #cfb274, #e4cb8f); color: #0a0c10; }
+[data-card-layout="strip"] .pg-card-strip[data-tier="silver"]   { background: linear-gradient(90deg, #a8b1bd, #c5cdd7); color: #0a0c10; }
+[data-card-layout="strip"] .pg-card-strip[data-tier="bronze"]   { background: linear-gradient(90deg, #b07d4f, #cf9e6f); color: #0a0c10; }
+[data-card-layout="strip"] .pg-card-strip[data-tier="borked"]   { background: linear-gradient(90deg, #884444, #b25555); color: #fff; }
+[data-card-layout="strip"] .pg-card-strip-tier {
+  font-size: 0.78rem;
+  font-weight: 800;
+  letter-spacing: 0.08em;
+  text-transform: uppercase;
+}
+[data-card-layout="strip"] .pg-card-strip .game-card-store-pill {
+  background: rgba(0, 0, 0, 0.22) !important;
+  color: inherit;
+  font-size: 0.62rem;
+  padding: 2px 8px;
+}
+[data-card-layout="strip"] .pg-right { display: none; }
 .pg-card:hover { border-color: var(--accent); }
 .pg-card:active { transform: translateY(1px); }
 .pg-thumb-wrap {
