+++ index.html
@@ -16,8 +16,8 @@
 <link rel="stylesheet" href="css/shared/topbar.css?v=f4f0510d">
 <link rel="stylesheet" href="css/shared/site.css?v=4d518a51">
 <link rel="stylesheet" href="css/shared/filters.css?v=26d5b7e1">
-<link rel="stylesheet" href="css/shared/cards.css?v=f95cc608">
-<link rel="stylesheet" href="css/index/index.css?v=d6d74d80">
+<link rel="stylesheet" href="css/shared/cards.css?v=5a77b933">
+<link rel="stylesheet" href="css/index/index.css?v=d47ad49c">
 </head>
 <body>
 
@@ -221,6 +221,6 @@
 <script src="js/lib/supabase-client.js?v=04813aa1"></script>
 <script src="js/lib/topbar.js?v=02244420"></script>
 <script src="js/lib/toast.js?v=7bb23014"></script>
-<script type="module" src="js/index/main.js?v=d53cf296"></script>
+<script type="module" src="js/index/main.js?v=7adace90"></script>
 </body>
 </html>
