+++ stats.html
@@ -13,7 +13,7 @@
 <link rel="stylesheet" href="css/shared/topbar.css?v=f4f0510d">
 <link rel="stylesheet" href="css/shared/site.css?v=4d518a51">
 <link rel="stylesheet" href="css/shared/filters.css?v=26d5b7e1">
-<link rel="stylesheet" href="css/shared/cards.css?v=f95cc608">
+<link rel="stylesheet" href="css/shared/cards.css?v=5a77b933">
 <style>
 h1 {
   font-family: var(--font-display);
