+++ terms.html
@@ -11,8 +11,8 @@
 <link rel="stylesheet" href="css/shared/base.css?v=45c7e06d">
 <link rel="stylesheet" href="css/shared/topbar.css?v=f4f0510d">
 <link rel="stylesheet" href="css/shared/site.css?v=4d518a51">
-<link rel="stylesheet" href="css/shared/cards.css?v=f95cc608">
-<link rel="stylesheet" href="css/index/index.css?v=d6d74d80">
+<link rel="stylesheet" href="css/shared/cards.css?v=5a77b933">
+<link rel="stylesheet" href="css/index/index.css?v=d47ad49c">
 <style>
   .prose { max-width: 720px; margin: 2rem auto; line-height: 1.7; }
   .prose h1 { font-size: 1.6rem; margin-bottom: 0.5rem; }
